refactor: extract graph to ExploreMetricItemGraph.tsx
